pyproject.toml
[tool.poetry]
name = "best-realty-stats-bot"
version = "0.1.0"
description = ""
authors = ["a.groza <a.groza.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "stats_bot"}]


[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = {version = "==20.0", extras = ["job_queue"]}
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
aiosqlite = "^0.20.0"
python-multipart = "^0.0.9"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
